requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==3.2.25
django-crispy-forms==2.0
django-filter==23.5
django-jazzmin==2.6.2
environ==1.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1
